requirements.txt
Django==1.10.6
django-import-export==1.0.1
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-redis==4.9.0
DjangoUeditor==1.8.143
future==0.16.0
gunicorn==19.9.0
httplib2==0.9.2
jieba==0.39
Jinja2==2.8
Pillow==5.2.0
PyMySQL==0.9.2
redis==2.10.6
reload==0.9
requests==2.19.1
uWSGI==2.0.17.1
Whoosh==2.7.4
django-celery


